Refactored intersection elements so that they will accept either Immutable or Mutable inputs. I'm uneasy about not having an immutable option because of the relative weight of the signal and listener logic, but I have no use for immutable intersection objects right now so I'll add them back in when I need to.
I've added a mutable circle that accepts a radial point rather than a radius.

Completed the functionality that finds the points of intersection between a line and a polygon. There
is a problem with this: if there are n edges in the polygon then there are n potential intersections but only m (0 <= m < n) actual intersections. As the number of intersections change, how should these be referenced with respect to potential dependents... More thought is needed.

Refactored the drawers to be useful classes for composing graphical representations of the objects. Added draggable points so that mutability functionality can be visually tested. Created a test class 'src/examples/ThreePoints.as' to show off the functionality of the Drawers.