Added the Intersection of CircleAndLine
Altered the location of the line-multiplier check into the LineType so that it is availalbe in any context where there is a line. (In this case, for the circle-line intersection).

Added a VertexDrawer for drawing constrained vertices such as intersections

Added two example files which act as UI tests for specific intersection classes

Added RobotLegs for future development