The intersection of two mutable polygons turns out to be harder than you think! This is unfinished, but it satisfies the first unit test by correctly determining the intersection between two squares.
There is an unacknowledged assumption in the code that the polygons have the same direction; ie both are clockwise or anti-clockwise. This needs to be addressed.

So far the class only handles intersections which are calculated at the time when they occur; for object coherence this needs to be extended to calculating the change of positions of vertices and creating polygons the number of vertices in which can change as the intersection changes.