Edded a new example: IntersectTwoQuadrilaterals to act as a visual test case while I continue to work through the implications of polygon-polygon intersection.
I've made a mistake somewhere with respect to the intersection of two polygons with reverse directions, but I am committing this anyway, despite that unit-test failing.

On reflection, the best strategy to ensure intersection persistence when the definien polygons change is to add 'metadata' to the IntersectionOfTwoPolygonsVertex classes, which allows each vertex to identify itself with a particular intersection. If it remains NaN or remains real necessarily it remains part of the same polygon, but as its reality changes the intersection can be accordingly adjusted.

A polygon which accepts the addition and removal of vertices is required for this, which is an extension which may need to be pushed into the MutablePolygon also, in which casse this /twopolygons/ library may be simplifiable.