What I consider to be a fault with the Vector.<T> class has left me in a bind, so I have refactored out the Vectors in my library in favour of Array until the issue is settled. The problem is that a Vector.<ClassThatImplementsVertex> cannot be cast as a Vector.<Vertex>; they are considered incompatible types. This being despite the fact that the two classes are functionally equivalent and that every ClassThatImplementsVertex member of the Vector.<ClassThatImplementsVertex> is a Vertex. This is classified as "Not A Bug" by Adobe; I can foresee yet another disagreement with a know-it-all C developer from Adobe who read a manual about OOP once and thinks he's an expert. Humpf.
The intersection of two polygons code has been refactored out into multiple classes as it was becomming unreadable. The implementation of graceful handling of mutability is on-going, as it is a pain in the ass. This work has highlighted the need for additive collections. A polygon is a collection of vertices, and an intersection of polygons is a collection of polygons. In the case of intersecting two polygons, the collection of polygons may grow or srink in size, and the constituent polygons may similarly add or lose vertices. The resultant AdditiveCollection interface is an initial reaction to this problem.

The problem of polygon direction has been solved, at least to the satisfaction of the current unit-tests. Further testing of the UI structures that are created will eventually highlight corner-conditions exist that have not yet been considered. There probably are.

Alec