Refactor to update the as3-signals library and to reflect it's API changes.
There is currently a problem with initialisation and vertical dependencies. As complexities add onto each other it is unclear when vertically dependent elements should call their changed methods and when they should update themselves. I am thinking through a strategy for coping with this along the following lines:

• Add an Invalidates interface (this has already been done) which extends Mutable and is intended to be implemented for objects that have Mutable definiens, a change in which should not immediately necessitate a recalculation (if I could drag an entire polygon, all the polygons' vertices would dispatch changed signals, but a reliant object should not recalculate n times, for example).

• Create an Invalidation Manager which handles the dependency tree between elements and sequentially updates objects in the right order so that an object is recalculated only when all definien elements are already updated.

This leads to whether I should start to adopt RobotLegs now in order to inject the Invalidation manager into classes; it feels as though it should be a singleton, and RobotLegs offers the best mechanism for implementing Singletons.